+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/screens/NetworkSelection/components/index.tsx
@@ -6,6 +6,7 @@ import { useModularDrawerAnalytics } from "../../../analytics/useModularDrawerAn
 import { MODULAR_DRAWER_PAGE_NAME } from "../../../analytics/types";
 import { EnhancedModularDrawerConfiguration } from "@ledgerhq/live-common/wallet-api/ModularDrawer/types";
 import createNetworkConfigurationHook from "../modules/createNetworkConfigurationHook";
+import { CurrenciesByProviderId } from "@ledgerhq/live-common/deposit/type";
 
 type SelectNetworkProps = {
   networks?: CryptoOrTokenCurrency[];
@@ -13,6 +14,8 @@ type SelectNetworkProps = {
   flow: string;
   onNetworkSelected: (network: CryptoOrTokenCurrency) => void;
   networksConfig: EnhancedModularDrawerConfiguration["networks"];
+  currenciesByProvider: CurrenciesByProviderId[];
+  selectedAssetId?: string;
 };
 
 export const SelectNetwork = ({
@@ -21,15 +24,19 @@ export const SelectNetwork = ({
   source,
   flow,
   networksConfig,
+  currenciesByProvider,
+  selectedAssetId,
 }: SelectNetworkProps) => {
   const { trackModularDrawerEvent } = useModularDrawerAnalytics();
 
-  if (!networks || networks.length === 0) {
+  if (!networks || networks.length === 0 || !selectedAssetId) {
     return null;
   }
 
   const transformNetworks = createNetworkConfigurationHook({
     networksConfig,
+    currenciesByProvider,
+    selectedAssetId,
   });
 
   const formattedNetworks = transformNetworks(networks);
@@ -48,6 +55,7 @@ export const SelectNetwork = ({
       },
       {
         formatNetworkConfig: true,
+        networksConfig,
       },
     );