+++ libs/live-signer-solana/src/LegacySignerSolana.ts
@@ -14,11 +14,22 @@ import trustService from "@ledgerhq/ledger-trust-service";
 import { loadPKI } from "@ledgerhq/hw-bolos";
 import { LatestFirmwareVersionRequired, UpdateYourApp } from "@ledgerhq/errors";
 
-const TRUSTED_NAME_MIN_VERSION = "1.7.1";
+const TRUSTED_NAME_MIN_VERSION = "1.8.1";
+const DYNAMIC_DESCRIPTOR_MIN_VERSION = "1.9.0";
 const MANAGER_APP_NAME = "Solana";
 
 function isPKIUnsupportedError(err: unknown): err is TransportStatusError {
   return err instanceof TransportStatusError && err.message.includes("0x6a81");
+}
+
+async function tryLoadPKI(...args: Parameters<typeof loadPKI>) {
+  try {
+    await loadPKI(...args);
+  } catch (err) {
+    if (isPKIUnsupportedError(err)) {
+      throw new LatestFirmwareVersionRequired("LatestFirmwareVersionRequired");
+    }
+  }
 }
 
 export class LegacySignerSolana implements SolanaSigner {
@@ -34,13 +45,19 @@ export class LegacySignerSolana implements SolanaSigner {
     return this.signer.getAppConfiguration();
   }
 
-  private async checkAppVersion() {
+  private async checkAppVersion(
+    minVersion: string,
+    { throwOnOutdated }: { throwOnOutdated: boolean },
+  ) {
     const { version } = await this.getAppConfiguration();
-    if (semver.lt(version, TRUSTED_NAME_MIN_VERSION)) {
+    const outdated = semver.lt(version, minVersion);
+
+    if (outdated && throwOnOutdated) {
       throw new UpdateYourApp(undefined, {
         managerAppName: MANAGER_APP_NAME,
       });
     }
+    return !outdated;
   }
 
   getAddress(path: string, display?: boolean | undefined): Promise<SolanaAddress> {
@@ -60,6 +77,7 @@ export class LegacySignerSolana implements SolanaSigner {
       if (resolution.deviceModelId !== DeviceModelId.nanoS) {
         const { descriptor, signature } = await calService.getCertificate(
           resolution.deviceModelId,
+          "trusted_name",
           "latest",
           { signatureKind: resolution.certificateSignatureKind },
         );
@@ -73,7 +91,7 @@ export class LegacySignerSolana implements SolanaSigner {
         }
 
         if (resolution.tokenAddress) {
-          await this.checkAppVersion();
+          await this.checkAppVersion(TRUSTED_NAME_MIN_VERSION, { throwOnOutdated: true });
 
           const challenge = await this.signer.getChallenge();
           const { signedDescriptor } = await trustService.getOwnerAddress(
@@ -86,7 +104,7 @@ export class LegacySignerSolana implements SolanaSigner {
           }
         }
         if (resolution.createATA) {
-          await this.checkAppVersion();
+          await this.checkAppVersion(TRUSTED_NAME_MIN_VERSION, { throwOnOutdated: true });
 
           const challenge = await this.signer.getChallenge();
           const { signedDescriptor } = await trustService.computedTokenAddress(
@@ -98,6 +116,31 @@ export class LegacySignerSolana implements SolanaSigner {
           if (signedDescriptor) {
             await this.signer.provideTrustedName(signedDescriptor);
           }
+        }
+
+        const dynamicDescriptorSupport = await this.checkAppVersion(
+          DYNAMIC_DESCRIPTOR_MIN_VERSION,
+          {
+            throwOnOutdated: false,
+          },
+        );
+        if (dynamicDescriptorSupport && resolution.tokenInternalId) {
+          const { descriptor: coinMetaDescriptor, signature: coinMetaSignature } =
+            await calService.getCertificate(resolution.deviceModelId, "coin_meta", "latest", {
+              signatureKind: resolution.certificateSignatureKind,
+            });
+
+          await tryLoadPKI(this.transport, "COIN_META", coinMetaDescriptor, coinMetaSignature);
+
+          const token = await calService.findToken(
+            { id: resolution.tokenInternalId },
+            { signatureKind: resolution.certificateSignatureKind },
+          );
+
+          await this.signer.provideTrustedDynamicDescriptor({
+            data: Buffer.from(token.descriptor.data, "hex"),
+            signature: Buffer.from(token.descriptor.signature, "hex"),
+          });
         }
       }
     }