+++ libs/live-signer-solana/tests/LegacySignerSolana.test.ts
@@ -7,13 +7,19 @@ import {
   TransportStatusError,
   UpdateYourApp,
 } from "@ledgerhq/errors";
-import { DeviceModelId } from "@ledgerhq/devices/lib/index";
+import { DeviceModelId } from "@ledgerhq/devices/index";
 import calService from "@ledgerhq/ledger-cal-service";
 import trustService from "@ledgerhq/ledger-trust-service";
+import { TokenData } from "@ledgerhq/ledger-cal-service/tokens";
+import Transport from "@ledgerhq/hw-transport";
 
-const signer = new LegacySignerSolana({ decorateAppAPIMethods: () => {} } as any);
+const signer = new LegacySignerSolana({ decorateAppAPIMethods: () => {} } as unknown as Transport);
 
 describe("LegacySignerSolana", () => {
+  afterEach(() => {
+    jest.restoreAllMocks();
+  });
+
   describe("getAppConfiguration", () => {
     it("gets the app configuration", async () => {
       const getAppConfiguration = jest
@@ -84,7 +90,7 @@ describe("LegacySignerSolana", () => {
       jest.spyOn(Solana.prototype, "getAppConfiguration").mockResolvedValue({
         blindSigningEnabled: false,
         pubKeyDisplayMode: PubKeyDisplayMode.SHORT,
-        version: "1.6.2",
+        version: "1.7.1",
       });
 
       await expect(
@@ -137,9 +143,14 @@ describe("LegacySignerSolana", () => {
         ).toEqual({
           signature: Buffer.from("0102", "hex"),
         });
-        expect(getCertificate).toHaveBeenCalledWith(DeviceModelId.europa, "latest", {
-          signatureKind: expectedSignatureKind,
-        });
+        expect(getCertificate).toHaveBeenCalledWith(
+          DeviceModelId.europa,
+          "trusted_name",
+          "latest",
+          {
+            signatureKind: expectedSignatureKind,
+          },
+        );
         expect(loadPKI).toHaveBeenCalledWith(
           expect.anything(),
           "TRUSTED_NAME",
@@ -193,9 +204,14 @@ describe("LegacySignerSolana", () => {
         ).toEqual({
           signature: Buffer.from("0102", "hex"),
         });
-        expect(getCertificate).toHaveBeenCalledWith(DeviceModelId.europa, "latest", {
-          signatureKind: expectedSignatureKind,
-        });
+        expect(getCertificate).toHaveBeenCalledWith(
+          DeviceModelId.europa,
+          "trusted_name",
+          "latest",
+          {
+            signatureKind: expectedSignatureKind,
+          },
+        );
         expect(loadPKI).toHaveBeenCalledWith(
           expect.anything(),
           "TRUSTED_NAME",
@@ -204,6 +220,92 @@ describe("LegacySignerSolana", () => {
         );
         expect(computedTokenAddress).toHaveBeenCalledWith("address", "mintAddress", "challenge");
         expect(provideTrustedName).toHaveBeenCalledWith("signedDescriptor");
+        expect(signTransaction).toHaveBeenCalledWith("path", Buffer.from("transaction"));
+      },
+    );
+
+    it("does not try to register SPL info dynamically with app version prior to 1.9.0", async () => {
+      jest.spyOn(Solana.prototype, "getAppConfiguration").mockResolvedValue({
+        blindSigningEnabled: false,
+        pubKeyDisplayMode: PubKeyDisplayMode.SHORT,
+        version: "1.8.2",
+      });
+      jest.spyOn(calService, "getCertificate").mockResolvedValue({
+        descriptor: "certificateDescriptor",
+        signature: "certificateSignature",
+      });
+      const provideTrustedDynamicDescriptor = jest.spyOn(
+        Solana.prototype,
+        "provideTrustedDynamicDescriptor",
+      );
+      const signTransaction = jest.spyOn(Solana.prototype, "signTransaction").mockResolvedValue({
+        signature: Buffer.from("0102", "hex"),
+      });
+
+      expect(
+        await signer.signTransaction("path", Buffer.from("transaction"), {
+          deviceModelId: DeviceModelId.europa,
+          tokenInternalId: "tokenInternalId",
+        }),
+      ).toEqual({
+        signature: Buffer.from("0102", "hex"),
+      });
+      expect(provideTrustedDynamicDescriptor).not.toHaveBeenCalled();
+      expect(signTransaction).toHaveBeenCalledWith("path", Buffer.from("transaction"));
+    });
+
+    it.each([["prod"], ["test"]] as const)(
+      "signs a transaction with a resolution and a 'tokenInternalId' property, using the '%s' PKI",
+      async expectedSignatureKind => {
+        jest.spyOn(Solana.prototype, "getAppConfiguration").mockResolvedValue({
+          blindSigningEnabled: false,
+          pubKeyDisplayMode: PubKeyDisplayMode.SHORT,
+          version: "1.9.0",
+        });
+        const getCertificate = jest.spyOn(calService, "getCertificate").mockResolvedValue({
+          descriptor: "certificateDescriptor",
+          signature: "certificateSignature",
+        });
+        const loadPKI = jest.spyOn(loadPKIModule, "loadPKI").mockResolvedValue(undefined);
+        const findToken = jest.spyOn(calService, "findToken").mockResolvedValue({
+          descriptor: {
+            data: "0304",
+            signature: "0506",
+          },
+        } as unknown as TokenData);
+        const provideTrustedDynamicDescriptor = jest
+          .spyOn(Solana.prototype, "provideTrustedDynamicDescriptor")
+          .mockResolvedValue(true);
+        const signTransaction = jest.spyOn(Solana.prototype, "signTransaction").mockResolvedValue({
+          signature: Buffer.from("0102", "hex"),
+        });
+
+        expect(
+          await signer.signTransaction("path", Buffer.from("transaction"), {
+            deviceModelId: DeviceModelId.europa,
+            tokenInternalId: "tokenInternalId",
+            certificateSignatureKind: expectedSignatureKind,
+          }),
+        ).toEqual({
+          signature: Buffer.from("0102", "hex"),
+        });
+        expect(getCertificate).toHaveBeenCalledWith(DeviceModelId.europa, "coin_meta", "latest", {
+          signatureKind: expectedSignatureKind,
+        });
+        expect(loadPKI).toHaveBeenCalledWith(
+          expect.anything(),
+          "COIN_META",
+          "certificateDescriptor",
+          "certificateSignature",
+        );
+        expect(findToken).toHaveBeenCalledWith(
+          { id: "tokenInternalId" },
+          { signatureKind: expectedSignatureKind },
+        );
+        expect(provideTrustedDynamicDescriptor).toHaveBeenCalledWith({
+          data: Buffer.from("0304", "hex"),
+          signature: Buffer.from("0506", "hex"),
+        });
         expect(signTransaction).toHaveBeenCalledWith("path", Buffer.from("transaction"));
       },
     );