+++ libs/ledgerjs/packages/hw-app-celo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/hw-app-polkadot
 
+## 6.33.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.7
+
+## 6.33.5-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.7-next.0
+
 ## 6.33.4
 
 ### Patch Changes